Ta dificultando

+++ src/main/java/br/com/db1/dao/DAO.java
@@ -5,12 +5,12 @@ import java.util.List;
 public interface DAO<T> {
 	List<T> findAll();
 
-	T findById(Long id);
+	T findById(Integer id);
 
 	List<T> findByName(String name);
 
 	boolean save(T t);
 
-	boolean delete(Long id);
+	boolean delete(Integer id);
 
 }
